re-wrote redux and fetch using redux-toolkit
