+++ src/index.js
@@ -4,13 +4,7 @@ import "./index.css";
 import App from "./App";
 import { BrowserRouter as Router } from "react-router-dom";
 import { Provider} from 'react-redux'
-import { createStore, applyMiddleware } from "redux";
-import petsReducer from "./store/petSlice";
-import {composeWithDevTools} from 'redux-devtools-extension'
-import thunkMiddleware from 'redux-thunk'
-
-const composedEhancer = composeWithDevTools(applyMiddleware(thunkMiddleware))
-const store = createStore(petsReducer, composedEhancer)
+import store from './store/store'
 
 ReactDOM.render(
   <React.StrictMode>
