+++ src/store/petSlice.js
@@ -1,4 +1,10 @@
-import {createSlice} from "@reduxjs/toolkit"
+import {createAsyncThunk, createSlice} from "@reduxjs/toolkit"
+
+export const fetchPets = createAsyncThunk("pets/fetchPets",() => {
+  return fetch('/pets')
+  .then((response) => response.json())
+  .then(data =>  data)
+})
 
 const initialState = {
   petList: [],
@@ -16,9 +22,18 @@ const petSlice = createSlice({
       const pet = state.petList.find((cat) => cat.id === action.payload.id)
       pet.hungry = action.payload.hungry
     }
+  },
+  extraReducers: {
+    [fetchPets.pending](state){
+      state.status = "loading"
+    },
+    [fetchPets.fulfilled](state, action){
+      state.petList = action.payload
+      state.status = "idle"
+    }
   }
 })
 
-export const catActions = petSlice.actions
+export const petActions = petSlice.actions
 
 export default petSlice.reducer
